Remove info route from worker
To do this we ...

1.  Remove the baked in "info" link in the base template
2.  Add that to the scheduler's list of links
3.  Add a redirect from "info" to the actual page
4.  Create a generic redirect route
5.  Move that and the RequestHandler to utils to avoid code duplication
    between scheduler and worker

Fixes https://github.com/dask/distributed/issues/2722
