+++ distributed/dashboard/templates/base.html
@@ -29,9 +29,6 @@
           <a href="{{ page }}">{{ page|title }}</a>
         </li>
         {% endfor %}
-        <li>
-          <a href="info/main/workers.html">Info</a>
-        </li>
         <li id="navbar-toggle-icon">
           <a href="javascript:void(0);" onclick="myFunction()">
             <img src="statics/images/fa-bars.svg"></img>
