+++ distributed/dashboard/utils.py
@@ -1,13 +1,16 @@
 from __future__ import print_function, division, absolute_import
 
 from distutils.version import LooseVersion
+import os
 
 import bokeh
+from tornado import web
 from toolz import partition
 
 from ..compatibility import PY2
 
 BOKEH_VERSION = LooseVersion(bokeh.__version__)
+dirname = os.path.dirname(__file__)
 
 
 if BOKEH_VERSION >= "1.0.0" and not PY2:
@@ -32,3 +35,20 @@ def parse_args(args):
 def transpose(lod):
     keys = list(lod[0].keys())
     return {k: [d[k] for d in lod] for k in keys}
+
+
+class RequestHandler(web.RequestHandler):
+    def initialize(self, server=None, extra=None):
+        self.server = server
+        self.extra = extra or {}
+
+    def get_template_path(self):
+        return os.path.join(dirname, "templates")
+
+
+def redirect(path):
+    class Redirect(RequestHandler):
+        def get(self):
+            self.redirect(path)
+
+    return Redirect
